requirements.txt
PyYAML
click
backports.functools-lru-cache
cloudshell-automation-api
cloudshell-logging